Move some tests to online-judge-tools/api-client
